# Backend/swagger.yaml
openapi: 3.0.0
info:
  title: User Management API
  description: API for managing user accounts and events
  version: 1.0.0
servers:
  - url: /api/v1
    description: Main API server
security:
  - bearerAuth: []
paths:
...
  /users:
    post:
      tags:
        - Users
      summary: Create a new user
      description: Register a new user account
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Bad request (invalid input)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Users
      summary: Get all users
      description: Retrieve a paginated list of all users with optional filtering
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/sortBy'
        - $ref: '#/components/parameters/sortOrder'
        - name: role
          in: query
          description: Filter by user role
          schema:
            type: string
            enum: [Admin, Organizer, Participant]
        - name: status
          in: query
          description: Filter by user status
          schema:
            type: string
        - name: username
          in: query
          description: Filter by username (contains)
          schema:
            type: string
        - name: email
          in: query
          description: Filter by email (contains)
          schema:
            type: string
      responses:
        '200':
          description: List of users retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '404':
          description: No users found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{id}:
    get:
      tags:
        - Users
      summary: Get user by ID
      description: Retrieve details of a specific user
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/userId'
      responses:
        '200':
          description: User details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Users
      summary: Update user by ID
      description: Update details of an existing user
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/userId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Users
      summary: Delete user by ID
      description: Permanently delete a user account
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/userId'
      responses:
        '204':
          description: User deleted successfully
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/login:
    post:
      tags:
        - Users
      summary: Login user and return JWT token
      description: Authenticates a user using username and password, and returns a JWT token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  example: adminUser
                password:
                  type: string
                  format: password
                  example: securePassword123
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Login successful
                  token:
                    type: string
                    description: JWT token
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                        example: abc123
                      username:
                        type: string
                        example: adminUser
                      email:
                        type: string
                        example: admin@example.com
                      role:
                        type: string
                        example: admin
        '400':
          description: Missing username or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Event APIs
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
components:
  parameters:
    userId:
      name: id
      in: path
      required: true
      description: ID of the user
      schema:
        type: string
        example: xWMyZ4YEGZ
...
    page:
      name: page
      in: query
      description: Page number
      schema:
        type: integer
        default: 1
        minimum: 1
    limit:
      name: limit
      in: query
      description: Items per page
      schema:
        type: integer
        default: 10
        minimum: 1
        maximum: 100
    sortBy:
      name: sortBy
      in: query
      description: Sort by field
      schema:
        type: string
        enum: [createdAt, updatedAt, username, email]
        default: createdAt
    sortOrder:
      name: sortOrder
      in: query
      description: Sort direction
      schema:
        type: string
        enum: [asc, desc]
        default: desc
  schemas:
    UserCreate:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          example: johndoe
        email:
          type: string
          format: email
          example: john@example.com
        password:
          type: string
          format: password
          example: securePassword123
        role:
          type: string
          enum: [Admin, Organizer, Participant]
          default: Participant
        profile:
          type: object
          properties:
            firstName:
              type: string
              example: John
            lastName:
              type: string
              example: Doe
        status:
          type: string
          default: active
    UserUpdate:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [Admin, Organizer, Participant]
        profile:
          type: object
          properties:
            firstName:
              type: string
            lastName:
              type: string
    UserResponse:
      type: object
      properties:
        objectId:
          type: string
        username:
          type: string
        email:
          type: string
        role:
          type: string
        profile:
          type: object
          properties:
            firstName:
              type: string
            lastName:
              type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
        pagination:
          type: object
          properties:
            total:
              type: integer
            page:
              type: integer
            limit:
              type: integer
            totalPages:
              type: integer
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Error message describing the issue
    Error:
      type: object
      properties:
        error:
          type: string
          example: Invalid username or password
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT